python/requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.2.1
Django==5.2.5
Flask==3.1.2
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
six==1.17.0
sqlparse==0.5.3
Werkzeug==3.1.3